Minor re-org, moving all config related files into config folder
