+++ lib/config/index.js
@@ -2,11 +2,11 @@ const path = require('path');
 const fs = require('fs');
 const {EventEmitter} = require('events');
 const logger = require('winston');
-const CONSTANTS = require('./constants');
-const utils = require('./utils');
-const Theme = require('./theme');
 const merge = require('lodash/object/merge');
 const each = require('lodash/collection/each');
+const CONSTANTS = require('../constants');
+const utils = require('../utils');
+const Theme = require('./theme');
 
 /**
  * Object of event names assigned to an object for easy reference.
@@ -100,7 +100,7 @@ class Config extends EventEmitter {
    */
   static defaultConfig() {
     return utils.yaml.parse(
-      fs.readFileSync(path.resolve(__dirname, 'config_defaults.yml'), 'utf8')
+      fs.readFileSync(path.resolve(__dirname, 'defaults.yml'), 'utf8')
     );
   }
 }
