+++ lib/config/theme.js
@@ -1,10 +1,10 @@
 const glob = require('glob');
 const path = require('path');
 const fs = require('fs');
-const utils = require('./utils');
-const CONSTANTS = require('./constants');
 const merge = require('lodash/object/merge');
 const each = require('lodash/collection/each');
+const utils = require('../utils');
+const CONSTANTS = require('../constants');
 
 class Theme {
   constructor() {
